requirements.txt
fastapi
uvicorn[standard]
aiofiles
python-dotenv
openai
playwright
httpx
Jinja2
python-multipart

